Working Indexer CLI
