+++ packages/iobio-react-components/package.json
@@ -8,7 +8,7 @@
 	"scripts": {
 		"dev": "vite",
 		"build": "tsc -b && vite build",
-		"indexer": "pnpm exec tsx --env-file=.env src/utils/indexer.mts",
+		"indexer": "pnpm exec tsx --env-file=.env src/utils/launchIndexer.mts",
 		"lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
 		"preview": "vite preview",
 		"stats": "pnpm exec tsx --env-file=.env src/utils/launchStats.mts"
