--- packages/iobio-react-components/src/utils/indexer.mts
@@ -1,171 +1,0 @@
-/*
- *
- * Copyright (c) 2024 The Ontario Institute for Cancer Research. All rights reserved
- *
- *  This program and the accompanying materials are made available under the terms of
- *  the GNU Affero General Public License v3.0. You should have received a copy of the
- *  GNU Affero General Public License along with this program.
- *   If not, see <http://www.gnu.org/licenses/>.
- *
- *  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY
- *  EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
- *  OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT
- *  SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
- *  INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED
- *  TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
- *  OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER
- *  IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
- *  ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- *
- */
-
-import readline from 'node:readline/promises';
-import { BamFileExtensions } from './constants';
-import { getFileMetadata } from './scoreFileHelpers.mts';
-import { type ElasticSearchResult } from './scoreFileTypes';
-import { generateIobioStats } from './stats.mts';
-
-// Config
-const authKey = process.env.ES_AUTH_KEY;
-const esHost = process.env.ES_HOST_URL;
-if (!(authKey && esHost)) throw new Error('Required .env configuration values are missing');
-
-// Definitions
-const integer = 'integer';
-const float = 'float';
-const iobioProperties = JSON.stringify({
-	properties: {
-		iobio_metadata: {
-			properties: {
-				mapped_reads: { type: integer },
-				mapped_reads_percentage: { type: float },
-				forward_strands: { type: integer },
-				forward_strands_percentage: { type: float },
-				proper_pairs: { type: integer },
-				proper_pairs_percentage: { type: float },
-				singletons: { type: integer },
-				singletons_percentage: { type: float },
-				both_mates_mapped: { type: integer },
-				both_mates_mapped_percentage: { type: float },
-				duplicates: { type: integer },
-				duplicates_percentage: { type: float },
-				failed_qc: { type: integer },
-				first_mates: { type: integer },
-				last_read_position: { type: integer },
-				paired_end_reads: { type: integer },
-				reverse_strands: { type: integer },
-				second_mates: { type: integer },
-				total_reads: { type: integer },
-				mean_read_coverage: { type: integer },
-			},
-		},
-	},
-});
-
-// Script Start
-console.log('***** Overture Components: Iobio Metadata ElasticSearch Indexer *****');
-
-// User Input
-const readlineInterface = readline.createInterface({
-	input: process.stdin,
-	output: process.stdout,
-});
-
-const index = await readlineInterface.question('\nElasticSearch Index: ');
-const documentId = await readlineInterface.question('\nDocument Id: ');
-readlineInterface.close();
-if (!(index && documentId)) throw new Error('ElasticSearch index and documentId are required');
-
-// Validate Index & Document exist
-const requestOptions = {
-	headers: {
-		Authorization: `ApiKey ${authKey}`,
-	},
-};
-
-const mappingUrl = new URL(`${index}/_mapping`, esHost);
-const indexResponse = await fetch(mappingUrl, requestOptions).then((response) => response.json());
-const indexProperties = indexResponse[index]?.mappings.properties;
-if (!indexProperties) throw new Error(`Error retrieving field mapping for ElasticSearch index ${index}`);
-
-const searchUrl = new URL(`${index}/_search`, esHost);
-const searchQuery = JSON.stringify({
-	query: {
-		simple_query_string: {
-			query: documentId,
-			fields: ['object_id'],
-		},
-	},
-});
-
-const searchRequestOptions = {
-	headers: {
-		'Content-Type': 'application/json',
-		...requestOptions.headers,
-	},
-	method: 'POST',
-	body: searchQuery,
-};
-
-const searchResponse = await fetch(searchUrl, searchRequestOptions).then((response) => response.json());
-const searchResult: ElasticSearchResult = searchResponse.hits.hits[0];
-
-if (searchResult === undefined) {
-	throw new Error(`No document found with id ${documentId}`);
-}
-
-// File Handling
-const elasticDocument = searchResult._source;
-if (elasticDocument.file_type && !BamFileExtensions.includes(elasticDocument.file_type)) {
-	throw new Error(`File is not a BAM or CRAM file, found extension ${elasticDocument.file_type}`);
-}
-
-const fileName = elasticDocument.file?.name;
-const { fileMetadata, indexFileMetadata } = await getFileMetadata(elasticDocument);
-const fileUrl = fileMetadata?.parts[0]?.url || null;
-if (!fileUrl) {
-	throw new Error(`Unable to retrieve Score File URL for document with id: ${documentId}`);
-}
-const indexFileUrl = indexFileMetadata?.parts[0]?.url || null;
-
-// Update Index Mapping
-const hasIobioMapping = indexProperties.hasOwnProperty('iobio_metadata');
-if (!hasIobioMapping) {
-	console.log(`Updating Index ${index}`);
-
-	const updateMappingRequestOptions = {
-		headers: {
-			'Content-Type': 'application/json',
-			...requestOptions.headers,
-		},
-		method: 'PUT',
-		body: iobioProperties,
-	};
-
-	await fetch(mappingUrl, updateMappingRequestOptions).then((response) => response.json());
-}
-
-// Generate Stats & Update Document
-const iobio_metadata = await generateIobioStats({
-	fileUrl,
-	fileName,
-	indexFileUrl,
-	enableFileOutput: true,
-});
-
-const updateUrl = new URL(`${index}/_doc/${documentId}`, esHost);
-const updateBody = JSON.stringify({
-	iobio_metadata,
-});
-
-const updateDocumentRequestOptions = {
-	headers: {
-		'Content-Type': 'application/json',
-		...requestOptions.headers,
-	},
-	method: 'POST',
-	body: updateBody,
-};
-
-const updateResponse = await fetch(updateUrl, updateDocumentRequestOptions).then((response) => response.json());
-console.log('updateResponse', updateResponse);
